requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.5
mediapipe>=0.8.10
tqdm>=4.62.0
flask>=2.0.0
pillow>=9.0.0
scikit-learn>=1.0.0